Updated postman collections, tested three main endpoints user, account, role. Validations completed for each endpoint. Added enpoint descriptions in README. Refactored Exception manager. Added Aspect for logging exceptions.Removed querybuilder. Added security handler for encryption, hashing, and jwt. Ran Sonar lint on all files.

+++ gamesgalore/src/main/java/com/revature/gamesgalore/serviceimpl/RoleServiceImpl.java
@@ -2,18 +2,20 @@ package com.revature.gamesgalore.serviceimpl;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.util.Collection;
+import java.util.List;
+
+import javax.transaction.Transactional;
 
 import com.revature.gamesgalore.dao.Role;
-import com.revature.gamesgalore.dao.entitydetails.RoleDetails;
 import com.revature.gamesgalore.exceptions.ExceptionManager;
 import com.revature.gamesgalore.repositories.RoleRepository;
 import com.revature.gamesgalore.service.RoleService;
-import com.revature.gamesgalore.util.QueryBuilder;
 
+@Transactional
 @Service
 public class RoleServiceImpl implements RoleService {
 
@@ -23,83 +25,87 @@ public class RoleServiceImpl implements RoleService {
 	RoleRepository roleRepository;
 
 	@Override
-	public Collection<Role> getRolesByParams(String roleName) {
+	public List<Role> getRolesByParams(String roleName) {
 		try {
-			QueryBuilder queryBuilder = new QueryBuilder();
-			queryBuilder.getSelectAll(RoleDetails.TABLE_NAME);
 			if (roleName != null) {
 				return roleRepository.findByRoleName(roleName);
 			}
 			return roleRepository.findAll();
 		} catch (Exception e) {
 			logger.error(e);
-			throw ExceptionManager.supplierThrows500Exception().get();
+			throw ExceptionManager.getRSE(HttpStatus.INTERNAL_SERVER_ERROR, ExceptionManager.UNEXPECTED_ERROR).get();
 		}
 	}
 
 	@Override
-	public void addRoles(Collection<Role> roles) {
+	public void addRoles(List<Role> roles) {
 		try {
 			for (Role r : roles) {
 				if (!validateRole(r)) {
-					throw ExceptionManager.supplierThrows400Exception().get();
+					throw ExceptionManager.getRSE(HttpStatus.BAD_REQUEST, ExceptionManager.VALIDATION_FAILED).get();
 				}
 				roleRepository.save(r);
 			}
 		} catch (ResponseStatusException rse) {
-			throw ExceptionManager.supplierThrows400Exception().get();
+			throw rse;
 		} catch (Exception e) {
 			logger.error(e);
-			throw ExceptionManager.supplierThrows500Exception().get();
+			throw ExceptionManager.getRSE(HttpStatus.INTERNAL_SERVER_ERROR, ExceptionManager.UNEXPECTED_ERROR).get();
 		}
-	}
-
-	@Override
-	public boolean validateRole(Role role) {
-		return !(role.getRoleName() == null || role.getRoleName().isEmpty());
 	}
 
 	@Override
 	public Role getRole(Long roleId) {
 		try {
-			return roleRepository.findById(roleId).orElseThrow(ExceptionManager.supplierThrows404Exception());
+			return roleRepository.findById(roleId).orElseThrow(ExceptionManager.getRSE(HttpStatus.NOT_FOUND, ExceptionManager.NOT_FOUND));
 		} catch (ResponseStatusException rse) {
-			throw ExceptionManager.supplierThrows404Exception().get();
+			throw rse;
 		} catch (Exception e) {
 			logger.error(e);
-			throw ExceptionManager.supplierThrows500Exception().get();
+			throw ExceptionManager.getRSE(HttpStatus.INTERNAL_SERVER_ERROR, ExceptionManager.UNEXPECTED_ERROR).get();
 		}
 	}
 
 	@Override
 	public void updateRole(Role role, Long roleId) {
 		try {
-			Role roleRetreived = roleRepository.findById(roleId).orElseThrow(ExceptionManager.supplierThrows404Exception());
+			Role roleRetreived = roleRepository.findById(roleId).orElseThrow(ExceptionManager.getRSE(HttpStatus.NOT_FOUND, ExceptionManager.NOT_FOUND));
 			setOverrides(roleRetreived,role);
+			if (!validateRole(roleRetreived)) {
+				throw ExceptionManager.getRSE(HttpStatus.BAD_REQUEST, ExceptionManager.VALIDATION_FAILED).get();
+			}
 			roleRepository.save(roleRetreived);
 		} catch (ResponseStatusException rse) {
-			throw ExceptionManager.supplierThrows404Exception().get();
+			throw rse;
 		} catch (Exception e) {
 			logger.error(e);
-			throw ExceptionManager.supplierThrows500Exception().get();
+			throw ExceptionManager.getRSE(HttpStatus.INTERNAL_SERVER_ERROR, ExceptionManager.UNEXPECTED_ERROR).get();
 		}
-	}
-
-	private void setOverrides(Role roleRetreived, Role role) {
-		if(role.getRoleName() != null) {roleRetreived.setRoleName(role.getRoleName());}
 	}
 	
 	@Override
 	public void deleteRole(Long roleId) {
 		try {
-			roleRepository.findById(roleId).orElseThrow(ExceptionManager.supplierThrows404Exception());
+			if(!roleRepository.findById(roleId).isPresent()) {
+				throw ExceptionManager.getRSE(HttpStatus.NOT_FOUND, ExceptionManager.NOT_FOUND).get();
+			}
 			roleRepository.deleteById(roleId);
 		} catch (ResponseStatusException rse) {
-			throw ExceptionManager.supplierThrows404Exception().get();
+			throw rse;
 		} catch (Exception e) {
 			logger.error(e);
-			throw ExceptionManager.supplierThrows500Exception().get();
+			throw ExceptionManager.getRSE(HttpStatus.INTERNAL_SERVER_ERROR, ExceptionManager.UNEXPECTED_ERROR).get();
 		}
+	}
+	
+	@Override
+	public boolean validateRole(Role role) {
+		return role.getRoleName() != null && !role.getRoleName().isEmpty();
+	}
+	
+	@Override
+	public void setOverrides(Role roleRetreived, Role role) {
+		if(role.getRoleName() != null) {roleRetreived.setRoleName(role.getRoleName());}
 	}
 
 }
